RedditCheck/requirements.txt
APScheduler==3.9.0
Django==4.2.2
djangorestframework==3.14.0
Pillow==9.5.0
pytest-playwright
